+++ tests/testthat/test-ztnhppp_agreement.R
@@ -1,4 +1,5 @@
 test_that("ztppp agrees with ppp for high intensities", {
+  skip_on_cran()
   set.seed(123)
   # High intensity, there are practically no zero event trajectories
   r_ppp_t <- unlist(lapply(integer(10000), function(x) ppp(t_min = 1, t_max = 3, rate = 10, atmost1 = TRUE)))
@@ -7,6 +8,7 @@ test_that("ztppp agrees with ppp for high intensities", {
 })
 
 test_that("ztppp agrees with ppp for low intensities", {
+  skip_on_cran()
   set.seed(123)
   # Low intensity, there are many zero event trajectories -- condition on having >=1 event
   r_ppp_t_low <- unlist(lapply(integer(10000), function(x) ppp(t_min = 1, t_max = 3, rate = 0.1, atmost1 = TRUE)))
@@ -15,6 +17,7 @@ test_that("ztppp agrees with ppp for low intensities", {
 })
 
 test_that("ztNHPPP methods agree on the first time to event with constant rate", {
+  skip_on_cran()
   set.seed(123)
   l <- function(t) rep(2, length(t))
   L <- function(t) 2 * t
@@ -38,6 +41,7 @@ test_that("ztNHPPP methods agree on the first time to event with constant rate",
 })
 
 test_that("ztNHPPP linear intensity agrees with general methods", {
+  skip_on_cran()
   set.seed(123)
   l <- function(t, intercept = 1, slope = 2) intercept + slope * t
   L <- function(t, intercept = 1, slope = 2, t0 = 1) Lambda_linear_form(t, intercept = intercept, slope = slope, t0 = t0)
@@ -53,6 +57,7 @@ test_that("ztNHPPP linear intensity agrees with general methods", {
 
 
 test_that("ztNHPPP linear intensity agrees with linear intensity for high rates", {
+  skip_on_cran()
   set.seed(123)
   r_ztnhppp_intens_linear <- unlist(lapply(integer(10000), function(x) ztdraw_sc_linear(intercept = 1, slope = 2, t_min = 1, t_max = 13, atmost1 = TRUE)))
   r_nhppp_intens_linear <- unlist(lapply(integer(10000), function(x) draw_sc_linear(intercept = 1, slope = 2, t_min = 1, t_max = 13, atmost1 = TRUE)))
@@ -60,6 +65,7 @@ test_that("ztNHPPP linear intensity agrees with linear intensity for high rates"
 })
 
 test_that("ztNHPPP linear intensity agrees with linear intensity for low rates", {
+  skip_on_cran()
   set.seed(123)
   r_nhppp_intens_linear <- unlist(lapply(integer(10000), function(x) draw_sc_linear(intercept = 1, slope = 2, t_min = 1, t_max = 1.1, atmost1 = TRUE)))
   r_ztnhppp_intens_linear <- unlist(lapply(integer(length(r_nhppp_intens_linear)), function(x) ztdraw_sc_linear(intercept = 1, slope = 2, t_min = 1, t_max = 1.1, atmost1 = TRUE)))
@@ -68,6 +74,7 @@ test_that("ztNHPPP linear intensity agrees with linear intensity for low rates",
 
 
 test_that("vztdraw_sc_step_regular() agrees with vdraw_sc_step_regular()", {
+  skip_on_cran()
   set.seed(123)
   Lmat <- matrix(rep(c(1, 11, 14, 17), 10000), ncol = 4, byrow = TRUE)
   r_vdraw_sc_step_regular <- vdraw_sc_step_regular(Lambda_matrix = Lmat, rate_matrix_t_min = 1, rate_matrix_t_max = 5, atmost1 = FALSE)
@@ -85,6 +92,7 @@ test_that("vztdraw_sc_step_regular() agrees with vdraw_sc_step_regular()", {
 
 
 test_that("vztdraw_sc_step_regular_cpp() agrees with vdraw_sc_step_regular()", {
+  skip_on_cran()
   set.seed(123)
   Lmat <- matrix(rep(c(1, 11, 14, 17), 10000), ncol = 4, byrow = TRUE)
   r_vdraw_sc_step_regular <- vdraw_sc_step_regular(Lambda_matrix = Lmat, rate_matrix_t_min = 1, rate_matrix_t_max = 5, atmost1 = FALSE)
@@ -103,6 +111,7 @@ test_that("vztdraw_sc_step_regular_cpp() agrees with vdraw_sc_step_regular()", {
 
 
 test_that("ztNHPPP exponential agrees with general methods", {
+  skip_on_cran()
   set.seed(123)
   l <- function(t, intercept = .1, slope = .02) exp(intercept + slope * t)
   L <- function(t, intercept = .1, slope = .02, t0 = 1) Lambda_exp_form(t, intercept = intercept, slope = slope, t0 = t0)
@@ -118,6 +127,7 @@ test_that("ztNHPPP exponential agrees with general methods", {
 
 
 test_that("ztNHPPP exponential agrees with NHPPP exponential for high rates", {
+  skip_on_cran()
   set.seed(123)
   r_nhppp_intens_exp <- unlist(lapply(integer(10000), function(x) draw_sc_loglinear(intercept = .1, slope = .02, t_min = 1, t_max = 13, atmost1 = TRUE)))
   r_ztnhppp_intens_exp <- unlist(lapply(integer(10000), function(x) ztdraw_sc_loglinear(intercept = .1, slope = .02, t_min = 1, t_max = 13, atmost1 = TRUE)))
@@ -126,6 +136,7 @@ test_that("ztNHPPP exponential agrees with NHPPP exponential for high rates", {
 
 
 test_that("ztNHPPP exponential agrees with NHPPP exponential for low rates", {
+  skip_on_cran()
   set.seed(123)
   r_nhppp_intens_exp <- unlist(lapply(integer(10000), function(x) draw_sc_loglinear(intercept = .1, slope = .02, t_min = 1, t_max = 1.1, atmost1 = TRUE)))
   r_ztnhppp_intens_exp <- unlist(lapply(integer(length(r_nhppp_intens_exp)), function(x) ztdraw_sc_loglinear(intercept = .1, slope = .02, t_min = 1, t_max = 1.1, atmost1 = TRUE)))